backend/requirements.txt
black==23.7.0
fastapi==0.101.1
uvicorn==0.23.2
flake8==6.1.0
motor==3.2.0
python-decouple==3.8
